+++ NEWS.md
@@ -5,6 +5,7 @@
 * Fixed map bounding rectangle for infinite isometric maps
 * Fixed tile selection to not get removed when deleting (#3281)
 * Fixed custom types not being usable without opening a project (#3295)
+* Fixed parallax layer positions for other maps in a world
 * Scripting: Avoid possible crash due to garbage collection (#3290)
 * Scripting: Fixed missing null check in Tileset.loadFromImage and Tile.setImage
 * Scripting: Initialize tile layer size also when added as part of a group layer (#3291)